updated functionality to involve proper exports for build
